requirement.txt
alembic==1.6.5
aniso8601==7.0.0
appdirs==1.4.4
APScheduler==3.9.1.post1
asgiref==3.4.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
black==21.6b0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.1
colorama==0.4.4
dnspython==2.2.1
fastapi==0.66.0
FastAPI-SQLAlchemy==0.2.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
idna==3.4
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
promise==2.3
psycopg2==2.9.1
pydantic==1.8.2
PyJWT==2.1.0
pymongo==4.3.3
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
regex==2021.7.1
requests==2.28.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
SQLAlchemy==1.4.20
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
uvicorn==0.14.0
